streamline database + better meta extraction

+++ ndms/derivative.py
@@ -2,14 +2,14 @@ import numpy as np
 import numba as nb
 
 
-@nb.njit(nb.float32[:, :](nb.float32[:, :]), nogil=True)
+@nb.njit(nb.float64[:, :](nb.float64[:, :]), nogil=True)
 def velocity(seq):
     """
     :param [n_frames x dim]
     """
     n_frames = len(seq)
     dim = seq.shape[1]
-    V = np.empty((n_frames - 1, dim), dtype=np.float32)
+    V = np.empty((n_frames - 1, dim), dtype=np.float64)
     for t in range(n_frames - 1):
         a = seq[t]
         b = seq[t + 1]
